Part-A Category filter added
