+++ app.vue
@@ -1,5 +1,6 @@
 <template>
   <div>
+    <Navbar />
     <NuxtPage />
   </div>
 </template>
